+++ ocp/src/main/java/exo3/ocp/Manager.java
@@ -2,13 +2,13 @@ package exo3.ocp;
 
 public class Manager extends Employe{
 
-	private int nbrordres;
-	public Manager(int salaire, int annee_courant, int annee_entree,int nbrordres) {
+	private double nbrordres;
+	public Manager(double salaire, double annee_courant, double annee_entree,double nbrordres) {
 		super(salaire, annee_courant, annee_entree);
 		this.nbrordres=nbrordres;
 	}
 
-	public int salaire() {
+	public double salaire() {
 	   	  return super.salaire()+(100*nbrordres);
 	     }
 }
